+++ app/Http/Controllers/API/ProductController.php
@@ -33,7 +33,7 @@ class ProductController extends Controller
     {
         $request = $request->validated();
 
-        if($product = Product::findOrFail($validated["id_product"])) {
+        if($product = Product::find($request["id_product"])) {
             return response()->json([
                 "success" => true,
                 "data" => $product
@@ -65,14 +65,12 @@ class ProductController extends Controller
         ], Response::HTTP_BAD_REQUEST);
     }
 
-    public function edit(ProductEditRequest $request) : JsonResponse
+    public function update(ProductEditRequest $request) : JsonResponse
     {
         $validated = $request->validated();
 
-        if($product = Product::findOrFail($validated["id_product"])) {
-            $product->fill($validated);
-
-            if($product->save()) {
+        if($product = Product::find($validated["id_product"])) {
+            if($product->update($validated)) {
                 return response()->json([
                     "success"=> true
                 ], Response::HTTP_OK);
@@ -87,13 +85,13 @@ class ProductController extends Controller
     public function delete(ProductByIdRequest $request) : JsonResponse
     {
         $request = $request->validated();
-
-        $product = Product::findOrFail($request["id_product"]);
-
-        if($product->delete()) {
-            return response()->json([
-                "success"=> true
-            ], Response::HTTP_OK);
+        
+        if($product = Product::find($request["id_product"])) {
+            if($product->delete()) {
+                return response()->json([
+                    "success"=> true
+                ], Response::HTTP_OK);
+            }    
         }
 
         return response()->json([
